requirements.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==1.0.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
packaging==23.2
passlib==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==3.0.1
